fix: Removed deprecated "hass.helpers" usage
